Update README.md
Since it's confusing players, I'm taking the liberty of removing the out of date Load Order Library link from the list of links at the top, and I'm changing the link further down that pointed there to point at the Changelog for now instead. We can revert this when the Load Order Library dev finishes the rewrite of that site. I'll also put something in the wiki about this to document it for the players.